+++ airbyte-integrations/bases/standard-source-test/src/main/java/io/airbyte/integrations/standardtest/source/TestPythonSourceMain.java
@@ -57,10 +57,10 @@ public class TestPythonSourceMain {
     final String imageName = ns.getString("imageName");
     final String pythonContainerName = ns.getString("pythonContainerName");
 
-    PythonSourceStandardTest.IMAGE_NAME = imageName;
-    PythonSourceStandardTest.PYTHON_CONTAINER_NAME = pythonContainerName;
+    PythonSourceAcceptanceTest.IMAGE_NAME = imageName;
+    PythonSourceAcceptanceTest.PYTHON_CONTAINER_NAME = pythonContainerName;
 
-    TestRunner.runTestClass(PythonSourceStandardTest.class);
+    TestRunner.runTestClass(PythonSourceAcceptanceTest.class);
   }
 
 }
