+++ airbyte-integrations/connectors/source-clickhouse/src/test-integration/java/io/airbyte/integrations/io/airbyte/integration_tests/sources/ClickHouseSourceAcceptanceTest.java
@@ -29,9 +29,11 @@ import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.commons.resources.MoreResources;
-import io.airbyte.db.Database;
 import io.airbyte.db.Databases;
-import io.airbyte.integrations.standardtest.source.SourceStandardTest;
+import io.airbyte.db.jdbc.JdbcDatabase;
+import io.airbyte.db.jdbc.JdbcUtils;
+import io.airbyte.integrations.source.clickhouse.ClickHouseSource;
+import io.airbyte.integrations.standardtest.source.SourceAcceptanceTest;
 import io.airbyte.protocol.models.CatalogHelpers;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.ConfiguredAirbyteStream;
@@ -43,68 +45,19 @@ import io.airbyte.protocol.models.SyncMode;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
-import org.jooq.SQLDialect;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.utility.MountableFile;
+import org.testcontainers.containers.ClickHouseContainer;
 
-// todo (cgardens) - Sanity check that when configured for CDC that postgres performs like any other
-// incremental source. As we have more sources support CDC we will find a more reusable way of doing
-// this, but for now this is a solid sanity check.
-public class CdcPostgresSourceStandardTest extends SourceStandardTest {
+public class ClickHouseSourceAcceptanceTest extends SourceAcceptanceTest {
 
-  private static final String SLOT_NAME_BASE = "debezium_slot";
-  private static final String STREAM_NAME = "public.id_and_name";
-  private static final String STREAM_NAME2 = "public.starships";
-
-  private PostgreSQLContainer<?> container;
+  private ClickHouseContainer db;
   private JsonNode config;
-
-  @Override
-  protected void setup(TestDestinationEnv testEnv) throws Exception {
-    container = new PostgreSQLContainer<>("postgres:13-alpine")
-        .withCopyFileToContainer(MountableFile.forClasspathResource("postgresql.conf"), "/etc/postgresql/postgresql.conf")
-        .withCommand("postgres -c config_file=/etc/postgresql/postgresql.conf");
-    container.start();
-
-    config = Jsons.jsonNode(ImmutableMap.builder()
-        .put("host", container.getHost())
-        .put("port", container.getFirstMappedPort())
-        .put("database", container.getDatabaseName())
-        .put("username", container.getUsername())
-        .put("password", container.getPassword())
-        .put("replication_method", ImmutableMap.of("replication_slot", SLOT_NAME_BASE))
-        .build());
-
-    final Database database = Databases.createDatabase(
-        config.get("username").asText(),
-        config.get("password").asText(),
-        String.format("jdbc:postgresql://%s:%s/%s",
-            config.get("host").asText(),
-            config.get("port").asText(),
-            config.get("database").asText()),
-        "org.postgresql.Driver",
-        SQLDialect.POSTGRES);
-
-    database.query(ctx -> {
-      ctx.execute("SELECT pg_create_logical_replication_slot('" + SLOT_NAME_BASE + "', 'pgoutput');");
-      ctx.execute("CREATE TABLE id_and_name(id INTEGER, name VARCHAR(200));");
-      ctx.execute("INSERT INTO id_and_name (id, name) VALUES (1,'picard'),  (2, 'crusher'), (3, 'vash');");
-      ctx.execute("CREATE TABLE starships(id INTEGER, name VARCHAR(200));");
-      ctx.execute("INSERT INTO starships (id, name) VALUES (1,'enterprise-d'),  (2, 'defiant'), (3, 'yamato');");
-      return null;
-    });
-
-    database.close();
-  }
-
-  @Override
-  protected void tearDown(TestDestinationEnv testEnv) {
-    container.close();
-  }
+  private static final String STREAM_NAME = "id_and_name";
+  private static final String STREAM_NAME2 = "starships";
+  private static final String SCHEMA_NAME = "default";
 
   @Override
   protected String getImageName() {
-    return "airbyte/source-postgres:dev";
+    return "airbyte/source-clickhouse:dev";
   }
 
   @Override
@@ -125,7 +78,7 @@ public class CdcPostgresSourceStandardTest extends SourceStandardTest {
             .withCursorField(Lists.newArrayList("id"))
             .withDestinationSyncMode(DestinationSyncMode.APPEND)
             .withStream(CatalogHelpers.createAirbyteStream(
-                STREAM_NAME,
+                String.format("%s.%s", config.get("database").asText(), STREAM_NAME),
                 Field.of("id", JsonSchemaPrimitive.NUMBER),
                 Field.of("name", JsonSchemaPrimitive.STRING))
                 .withSupportedSyncModes(Lists.newArrayList(SyncMode.FULL_REFRESH, SyncMode.INCREMENTAL))),
@@ -134,10 +87,15 @@ public class CdcPostgresSourceStandardTest extends SourceStandardTest {
             .withCursorField(Lists.newArrayList("id"))
             .withDestinationSyncMode(DestinationSyncMode.APPEND)
             .withStream(CatalogHelpers.createAirbyteStream(
-                STREAM_NAME2,
+                String.format("%s.%s", config.get("database").asText(), STREAM_NAME2),
                 Field.of("id", JsonSchemaPrimitive.NUMBER),
                 Field.of("name", JsonSchemaPrimitive.STRING))
                 .withSupportedSyncModes(Lists.newArrayList(SyncMode.FULL_REFRESH, SyncMode.INCREMENTAL)))));
+  }
+
+  @Override
+  protected JsonNode getState() {
+    return Jsons.jsonNode(new HashMap<>());
   }
 
   @Override
@@ -146,8 +104,53 @@ public class CdcPostgresSourceStandardTest extends SourceStandardTest {
   }
 
   @Override
-  protected JsonNode getState() {
-    return Jsons.jsonNode(new HashMap<>());
+  protected void setup(TestDestinationEnv testEnv) throws Exception {
+    db = new ClickHouseContainer("yandex/clickhouse-server:21.3.10.1-alpine");
+    db.start();
+
+    config = Jsons.jsonNode(ImmutableMap.builder()
+        .put("host", db.getHost())
+        .put("port", db.getFirstMappedPort())
+        .put("database", SCHEMA_NAME)
+        .put("username", db.getUsername())
+        .put("password", db.getPassword())
+        .build());
+
+    JdbcDatabase database = Databases.createJdbcDatabase(
+        config.get("username").asText(),
+        config.get("password").asText(),
+        String.format("jdbc:clickhouse://%s:%s/%s",
+            config.get("host").asText(),
+            config.get("port").asText(),
+            config.get("database").asText()),
+        ClickHouseSource.DRIVER_CLASS);
+
+    final String table1 = JdbcUtils.getFullyQualifiedTableName(SCHEMA_NAME, STREAM_NAME);
+    final String createTable1 =
+        String.format("CREATE TABLE IF NOT EXISTS %s (id INTEGER, name VARCHAR(200)) ENGINE = TinyLog \n", table1);
+    final String table2 = JdbcUtils.getFullyQualifiedTableName(SCHEMA_NAME, STREAM_NAME2);
+    final String createTable2 =
+        String.format("CREATE TABLE IF NOT EXISTS %s (id INTEGER, name VARCHAR(200)) ENGINE = TinyLog \n", table2);
+    database.execute(connection -> {
+      connection.createStatement().execute(createTable1);
+      connection.createStatement().execute(createTable2);
+    });
+
+    String insertTestData = String.format("INSERT INTO %s (id, name) VALUES (1,'picard'),  (2, 'crusher'), (3, 'vash');\n", table1);
+    String insertTestData2 = String.format("INSERT INTO %s (id, name) VALUES (1,'enterprise-d'),  (2, 'defiant'), (3, 'yamato');\n", table2);
+    database.execute(connection -> {
+      connection.createStatement().execute(insertTestData);
+      connection.createStatement().execute(insertTestData2);
+    });
+
+    database.close();
+  }
+
+  @Override
+  protected void tearDown(TestDestinationEnv testEnv) {
+    db.close();
+    db.stop();
+
   }
 
 }
