+++ airbyte-integrations/connectors/source-jdbc/src/testFixtures/java/io/airbyte/integrations/source/jdbc/test/JdbcSourceAcceptanceTest.java
@@ -83,7 +83,7 @@ import org.junit.jupiter.api.Test;
 // 3. From the class that extends this one, implement a @AfterEach that cleans out the database
 // between each test.
 // 4. Then implement the abstract methods documented below.
-public abstract class JdbcSourceStandardTest {
+public abstract class JdbcSourceAcceptanceTest {
 
   public static String SCHEMA_NAME = "jdbc_integration_test1";
   public static String SCHEMA_NAME2 = "jdbc_integration_test2";
