+++ airbyte-integrations/connectors/source-mysql/src/test-integration/java/io/airbyte/integrations/source/mysql/MySqlSourceAcceptanceTest.java
@@ -22,18 +22,16 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.io.airbyte.integration_tests.sources;
+package io.airbyte.integrations.source.mysql;
 
 import com.fasterxml.jackson.databind.JsonNode;
 import com.google.common.collect.ImmutableMap;
 import com.google.common.collect.Lists;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.commons.resources.MoreResources;
+import io.airbyte.db.Database;
 import io.airbyte.db.Databases;
-import io.airbyte.db.jdbc.JdbcDatabase;
-import io.airbyte.db.jdbc.JdbcUtils;
-import io.airbyte.integrations.source.clickhouse.ClickHouseSource;
-import io.airbyte.integrations.standardtest.source.SourceStandardTest;
+import io.airbyte.integrations.standardtest.source.SourceAcceptanceTest;
 import io.airbyte.protocol.models.CatalogHelpers;
 import io.airbyte.protocol.models.ConfiguredAirbyteCatalog;
 import io.airbyte.protocol.models.ConfiguredAirbyteStream;
@@ -45,19 +43,59 @@ import io.airbyte.protocol.models.SyncMode;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
-import org.testcontainers.containers.ClickHouseContainer;
+import org.jooq.SQLDialect;
+import org.testcontainers.containers.MySQLContainer;
 
-public class ClickHouseSourceStandardTest extends SourceStandardTest {
+public class MySqlSourceAcceptanceTest extends SourceAcceptanceTest {
 
-  private ClickHouseContainer db;
-  private JsonNode config;
   private static final String STREAM_NAME = "id_and_name";
-  private static final String STREAM_NAME2 = "starships";
-  private static final String SCHEMA_NAME = "default";
+  private static final String STREAM_NAME2 = "public.starships";
+
+  private MySQLContainer<?> container;
+  private JsonNode config;
+
+  @Override
+  protected void setup(TestDestinationEnv testEnv) throws Exception {
+    container = new MySQLContainer<>("mysql:8.0");
+    container.start();
+
+    config = Jsons.jsonNode(ImmutableMap.builder()
+        .put("host", container.getHost())
+        .put("port", container.getFirstMappedPort())
+        .put("database", container.getDatabaseName())
+        .put("username", container.getUsername())
+        .put("password", container.getPassword())
+        .build());
+
+    final Database database = Databases.createDatabase(
+        config.get("username").asText(),
+        config.get("password").asText(),
+        String.format("jdbc:mysql://%s:%s/%s",
+            config.get("host").asText(),
+            config.get("port").asText(),
+            config.get("database").asText()),
+        "com.mysql.cj.jdbc.Driver",
+        SQLDialect.MYSQL);
+
+    database.query(ctx -> {
+      ctx.fetch("CREATE TABLE id_and_name(id INTEGER, name VARCHAR(200));");
+      ctx.fetch("INSERT INTO id_and_name (id, name) VALUES (1,'picard'),  (2, 'crusher'), (3, 'vash');");
+      ctx.fetch("CREATE TABLE starships(id INTEGER, name VARCHAR(200));");
+      ctx.fetch("INSERT INTO starships (id, name) VALUES (1,'enterprise-d'),  (2, 'defiant'), (3, 'yamato');");
+      return null;
+    });
+
+    database.close();
+  }
+
+  @Override
+  protected void tearDown(TestDestinationEnv testEnv) {
+    container.close();
+  }
 
   @Override
   protected String getImageName() {
-    return "airbyte/source-clickhouse:dev";
+    return "airbyte/source-mysql:dev";
   }
 
   @Override
@@ -94,63 +132,13 @@ public class ClickHouseSourceStandardTest extends SourceStandardTest {
   }
 
   @Override
-  protected JsonNode getState() {
-    return Jsons.jsonNode(new HashMap<>());
-  }
-
-  @Override
   protected List<String> getRegexTests() {
     return Collections.emptyList();
   }
 
   @Override
-  protected void setup(TestDestinationEnv testEnv) throws Exception {
-    db = new ClickHouseContainer("yandex/clickhouse-server:21.3.10.1-alpine");
-    db.start();
-
-    config = Jsons.jsonNode(ImmutableMap.builder()
-        .put("host", db.getHost())
-        .put("port", db.getFirstMappedPort())
-        .put("database", SCHEMA_NAME)
-        .put("username", db.getUsername())
-        .put("password", db.getPassword())
-        .build());
-
-    JdbcDatabase database = Databases.createJdbcDatabase(
-        config.get("username").asText(),
-        config.get("password").asText(),
-        String.format("jdbc:clickhouse://%s:%s/%s",
-            config.get("host").asText(),
-            config.get("port").asText(),
-            config.get("database").asText()),
-        ClickHouseSource.DRIVER_CLASS);
-
-    final String table1 = JdbcUtils.getFullyQualifiedTableName(SCHEMA_NAME, STREAM_NAME);
-    final String createTable1 =
-        String.format("CREATE TABLE IF NOT EXISTS %s (id INTEGER, name VARCHAR(200)) ENGINE = TinyLog \n", table1);
-    final String table2 = JdbcUtils.getFullyQualifiedTableName(SCHEMA_NAME, STREAM_NAME2);
-    final String createTable2 =
-        String.format("CREATE TABLE IF NOT EXISTS %s (id INTEGER, name VARCHAR(200)) ENGINE = TinyLog \n", table2);
-    database.execute(connection -> {
-      connection.createStatement().execute(createTable1);
-      connection.createStatement().execute(createTable2);
-    });
-
-    String insertTestData = String.format("INSERT INTO %s (id, name) VALUES (1,'picard'),  (2, 'crusher'), (3, 'vash');\n", table1);
-    String insertTestData2 = String.format("INSERT INTO %s (id, name) VALUES (1,'enterprise-d'),  (2, 'defiant'), (3, 'yamato');\n", table2);
-    database.execute(connection -> {
-      connection.createStatement().execute(insertTestData);
-      connection.createStatement().execute(insertTestData2);
-    });
-
-    database.close();
-  }
-
-  @Override
-  protected void tearDown(TestDestinationEnv testEnv) {
-    db.close();
-    db.stop();
-
+  protected JsonNode getState() {
+    return Jsons.jsonNode(new HashMap<>());
   }
 
 }
