+++ airbyte-integrations/connectors/source-redshift/src/test-integration/java/io/airbyte/integrations/io/airbyte/integration_tests/sources/RedshiftJdbcSourceAcceptanceTest.java
@@ -22,50 +22,32 @@
  * SOFTWARE.
  */
 
-package io.airbyte.integrations.source.postgres;
+package io.airbyte.integrations.io.airbyte.integration_tests.sources;
 
 import com.fasterxml.jackson.databind.JsonNode;
-import com.google.common.collect.ImmutableMap;
 import io.airbyte.commons.io.IOs;
 import io.airbyte.commons.json.Jsons;
 import io.airbyte.integrations.source.jdbc.AbstractJdbcSource;
-import io.airbyte.integrations.source.jdbc.test.JdbcSourceStandardTest;
-import io.airbyte.test.utils.PostgreSQLContainerHelper;
-import org.apache.commons.lang3.RandomStringUtils;
-import org.junit.jupiter.api.AfterAll;
-import org.junit.jupiter.api.BeforeAll;
+import io.airbyte.integrations.source.jdbc.test.JdbcSourceAcceptanceTest;
+import io.airbyte.integrations.source.redshift.RedshiftSource;
+import java.nio.file.Path;
+import java.sql.SQLException;
+import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
-import org.testcontainers.containers.PostgreSQLContainer;
-import org.testcontainers.utility.MountableFile;
 
-class PostgresJdbcStandardSourceTest extends JdbcSourceStandardTest {
-
-  private static PostgreSQLContainer<?> PSQL_DB;
+// Run as part of integration tests, instead of unit tests, because there is no test container for
+// Redshift.
+class RedshiftJdbcSourceAcceptanceTest extends JdbcSourceAcceptanceTest {
 
   private JsonNode config;
 
-  @BeforeAll
-  static void init() {
-    PSQL_DB = new PostgreSQLContainer<>("postgres:13-alpine");
-    PSQL_DB.start();
+  private static JsonNode getStaticConfig() {
+    return Jsons.deserialize(IOs.readFile(Path.of("secrets/config.json")));
   }
 
   @BeforeEach
   public void setup() throws Exception {
-    final String dbName = "db_" + RandomStringUtils.randomAlphabetic(10).toLowerCase();
-
-    config = Jsons.jsonNode(ImmutableMap.builder()
-        .put("host", PSQL_DB.getHost())
-        .put("port", PSQL_DB.getFirstMappedPort())
-        .put("database", dbName)
-        .put("username", PSQL_DB.getUsername())
-        .put("password", PSQL_DB.getPassword())
-        .put("ssl", false)
-        .build());
-
-    final String initScriptName = "init_" + dbName.concat(".sql");
-    final String tmpFilePath = IOs.writeFileToRandomTmpDir(initScriptName, "CREATE DATABASE " + dbName + ";");
-    PostgreSQLContainerHelper.runSqlScript(MountableFile.forHostPath(tmpFilePath), PSQL_DB);
+    config = getStaticConfig();
 
     super.setup();
   }
@@ -77,7 +59,7 @@ class PostgresJdbcStandardSourceTest extends JdbcSourceStandardTest {
 
   @Override
   public AbstractJdbcSource getSource() {
-    return new PostgresSource();
+    return new RedshiftSource();
   }
 
   @Override
@@ -87,12 +69,12 @@ class PostgresJdbcStandardSourceTest extends JdbcSourceStandardTest {
 
   @Override
   public String getDriverClass() {
-    return PostgresSource.DRIVER_CLASS;
+    return RedshiftSource.DRIVER_CLASS;
   }
 
-  @AfterAll
-  static void cleanUp() {
-    PSQL_DB.close();
+  @AfterEach
+  public void tearDownRedshift() throws SQLException {
+    super.tearDown();
   }
 
 }
